remove internalWriteFiles from todo
